[TEST] Integration test cases

+++ GPM.Cubes.Application/Contracts/ICubeAppService.cs
@@ -6,6 +6,6 @@ namespace GPM.Cubes.Application.Contracts
     {
         bool HaveCollision(CubeCoordinates cube, CubeCoordinates cube2);
 
-        float GetIntersection(CubeCoordinates cube1, CubeCoordinates cube2);
+        double GetIntersection(CubeCoordinates cube1, CubeCoordinates cube2);
     }
 }
