+++ GPM.Cubes.Domain/Entities/Vector.cs
@@ -2,11 +2,11 @@
 {
     public class Vector
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double z { get; set; }
 
-        public Vector(float x, float y, float z)
+        public Vector(double x, double y, double z)
         {
             this.x = x;
             this.y = y;
